% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cp_wrapper.R
\name{cp_wrapper}
\alias{cp_wrapper}
\title{Find change points in a sequence of choices based on given test and decision criterion}
\usage{
cp_wrapper(input, test, Crit)
}
\arguments{
\item{input}{A vector of correct or incorrect decisions, coded as 1s and 0s}

\item{test}{Name of the test to be performed. Currently the following tests are implemented: "binomial", "chisquare".}

\item{Crit}{A real-valued decision criterion on logit. Values recommended by Gallistel et al. (2004)
are between 1.3 and 6, which correspond to p values of 0.05 and 0.000001, respectively.
The values are the logarithms of the odds against the null (no-change) hypothesis.
logit = log[(1-p)/p], where p is the desired significance level}
}
\value{
A data frame with the number of Trials (Trial) for each change point detected,
the cumulative number of correct responses, or successes, (CumSs), and the ratio of correct/total choices (Slopes)
for the trials since the last change point and the current change point.
}
\description{
Find change points in a sequence of choices based on given test and decision criterion
}
\examples{
# generate dummy data with a change point at trial 120
input <- c(rbinom(120,1,0.5),rbinom(200,1,0.8))
cp_wrapper(input, "chisquare", 3)
cp_wrapper(input, "binomial", 3)

# use included eyeblink data set:
eyeblink[,] # inspect data set
cp_wrapper(eyeblink, "chisquare", 4)
# using small criterion, e.g. 2, fails due to computational problems
# better use "binomial" instead
cp_wrapper(eyeblink, "binomial", 2)

# use included plusmaze data set:
plusmaze[,] # inspect data set
(cp.1 <- cp_wrapper(plusmaze, "binomial", 1.3))
# decrease sensitivity and detect different change points
(cp.2 <- cp_wrapper(plusmaze, "binomial", 2))
# the chisquare test detects more change points even with higher criterion
(cp.3 <- cp_wrapper(plusmaze, "chisquare", 2))

# plotting data with ggplot2
library(ggplot2)
# first generate data.frame with cumulative responses:
my.data <- data.frame(Trial = 1:length(plusmaze[,]),
CumRespMeasure = cumsum(plusmaze)[,])
# plot cumulative learning curve with change points
ggplot(my.data) + geom_line(aes(Trial,CumRespMeasure)) +
 geom_point(data=cp.1, aes(Trial, CumSs), shape=1, size = 6, color = "blue") +
 geom_point(data=cp.2, aes(Trial, CumSs), shape=2, size = 3, color = "green") +
 geom_point(data=cp.3, aes(Trial, CumSs), shape=3, size = 3, color = "red")

# plot average response rate per trial, with dplyr::lead
ggplot() + geom_step(data=cp.1, aes(Trial,dplyr::lead(Slopes))) +
 ylab("Average Response Rate per Trial") # ignore Warning message
}

