% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpfinders.R
\name{cpc}
\alias{cpc}
\title{Find putative change point in continuous-time cumulative records}
\usage{
cpc(Cum)
}
\arguments{
\item{Cum}{Input vector of the cumulative interevent intervals}
}
\value{
R A vector with the same length as Cum, with putative change points for each
event. The putative change point corresponding to the Nth
event is the preceding event at which the deviation of the observed event count
from the expected event count is maximal. The expected event count
at any earlier event, n, is Cum[n], the interval up to the nth event,
divided by the average interevent interval over the range from n = 0 to n = N
The deviation from expectation is n - this expectation. R is the value of n at
which this deviation is maximal
}
\description{
Find putative change point in continuous-time cumulative records
}
\details{
Not normally called directly, but via the cp_wrapper function instead
}

