% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpfinders.R
\name{cpd}
\alias{cpd}
\title{Find putative change point in discrete-time cumulative records}
\usage{
cpd(Cum)
}
\arguments{
\item{Cum}{Input vector of cumulative response counts}
}
\value{
R A vector with the same length as Cum, with putative change points for each
trial. These correspond to the preceding trial at which the deviation of the
observed count from the expected count is maximal
}
\description{
Find putative change point in discrete-time cumulative records
}
\details{
Not normally called directly, but via the cp_wrapper function instead.
}
