% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdetectoR.R
\docType{package}
\name{cpdetectoR}
\alias{cpdetectoR}
\alias{cpdetectoR-package}
\title{cpdetectoR: Change point estimation in learning curves}
\description{
The code is from Gallistel et al. (2004), translated from Matlab to R, with large
portions of comments and original code preserved. The package consists of a few
internal functions and one wrapper function, cp_wrapper.
}
