% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpdetectorr.R
\docType{package}
\name{cpdetectorr}
\alias{cpdetectorr}
\alias{cpdetectorr-package}
\title{cpdetector: Change point estimation in learning curves.}
\description{
The code is adapted from Gallistel et al. (2004).
The package consists of a few internal functions and one wrapper function, cp_wrapper.
}

