% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logits.R
\name{cpt}
\alias{cpt}
\title{Uses t test to find first significant change point}
\usage{
cpt(Data, R, Crit)
}
\arguments{
\item{Data}{A vector of trial by trial measures or successive intervals}

\item{R}{A vector of putative change points}

\item{Crit}{Decision criterion, the value the logit must exceed
for the function to return a significant change point}
}
\value{
CP The first significant change point
}
\description{
Uses t test to find first significant change point
}
\details{
This test is appropriate if one is looking for a change in the expectation of
a renewal event-generating process, where the interevent intervals are normally (rather
than exponentially) distributed. Not normally called directly, but via the cp_wrapper
function instead.
}
