% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdetectoR.R
\docType{data}
\name{eyeblink}
\alias{eyeblink}
\title{Data on rabbit eyeblink conditioning}
\format{An object of class \code{data.frame} with 500 rows and 1 columns.}
\source{
\url{http://www.pnas.org/content/suppl/2004/08/31/0404965101.DC1/04965DataSet5.txt/}
}
\usage{
eyeblink
}
\description{
Data from Gallistel et al. (2004). Original data description: "[Data] are binary,
1 or 0, according as a conditioned blink did or did not occur on a trial. They are
discrete-trial data, so the cpd function must be used to find putative change points.
They are frequency data, so one might try using the chisquare test. However, these are
the kind of data that cause the chisquare test to run into computational difficulties.
The random rate test works better with these data."
}
\keyword{datasets}
