% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdetectorr.R
\docType{data}
\name{feedingpref}
\alias{feedingpref}
\title{Data on feeding-by-feeding preference scores in a mouse matching experiment}
\format{An object of class \code{data.frame} with 641 rows and 1 columns.}
\source{
\url{http://www.pnas.org/content/suppl/2004/08/31/0404965101.DC1/04965DataSet4.txt/}
}
\usage{
feedingpref
}
\description{
Data from Gallistel et al. (2004). Original data description: "The side-preference
score during any one interfeeding interval is the difference between the amounts of
time spent at each feeding hopper divided by their sum. These are discrete-trial,
real-valued measures [...] They do not obey any standard distribution,
so the Kolmogorov-Smirnov test is appropriate for comparing the distributions before
and after a putative change point."
}
\keyword{datasets}
