% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdetectorr.R
\docType{data}
\name{hopperentry}
\alias{hopperentry}
\title{Data on successive hopper-entry speeds}
\format{An object of class \code{data.frame} with 244 rows and 1 columns.}
\source{
\url{http://www.pnas.org/content/suppl/2004/08/31/0404965101.DC1/04965DataSet2.txt/}
}
\usage{
hopperentry
}
\description{
Data from Gallistel et al. (2004). Original data description: "A hopper entry speed is the
reciprocal of the latency between the rise of the grain bin into the feeding hopper and the
entry of the pigeon's head into the hopper. These data are an example of a discrete-time
data record. [...] The entry speeds are approximately normally distributed [... t]hus, the the t-test
is appropriate for testing for a change in the mean entry latency."
}
\keyword{datasets}

