% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logits.R
\name{KS}
\alias{KS}
\title{Uses Komolgorov-Smirnov test to find first significant change point}
\usage{
KS(Data, R, Crit)
}
\arguments{
\item{Data}{A vector of trial by trial measures or successive intervals}

\item{R}{A vector of putative change points}

\item{Crit}{Decision criterion, the value the logit must exceed
for the function to return a significant change point}
}
\value{
r1 the change point row when the decision criterion is exceeded
}
\description{
Uses Komolgorov-Smirnov test to find first significant change point
}
\details{
Not normally called directly, but via the cp_wrapper function instead.
}
