% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lag_v}
\alias{lag_v}
\title{Make a lagged copy of a vector}
\usage{
lag_v(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
y A lagged copy of the input vector x, with the first entry repeated
}
\description{
Make a lagged copy of a vector
}
\details{
Not normally called directly, but via the cp_wrapper function instead.
}
