% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lead_v}
\alias{lead_v}
\title{Make a lead copy of a vector}
\usage{
lead_v(x)
}
\arguments{
\item{x}{Input vector of cumulative response counts}
}
\value{
y A lead copy of the input vector x, with the last entry repeated
}
\description{
Make a lead copy of a vector
}
\details{
Not normally called directly, but via the cp_wrapper function instead.
}
