% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdetectorr.R
\docType{data}
\name{matching}
\alias{matching}
\title{Data on interreward intervals in a matching experiment with concurrent variable
interval schedules}
\format{An object of class \code{data.frame} with 600 rows and 1 columns.}
\source{
\url{http://www.pnas.org/content/suppl/2004/08/31/0404965101.DC1/04965DataSet1.txt/}
}
\usage{
matching
}
\description{
Data from Gallistel et al. (2004). Original data description: "These are an example of
a continuous-time data record: the successive entries are the durations of the
successive interreward intervals. Thus, the putative change points must be found by
[setting isDiscrete = FALSE]. The interevent intervals are approximately exponentially
distributed [...]. Thus, the process approximates a random rate process and the random
rate logit method is appropriate [...]."
}
\keyword{datasets}
