% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdetectoR.R
\docType{data}
\name{plusmaze}
\alias{plusmaze}
\title{Data on correct choices in a + maze}
\format{An object of class \code{data.frame} with 200 rows and 1 columns.}
\source{
\url{http://www.pnas.org/content/suppl/2004/08/31/0404965101.DC1/04965DataSet3.txt/}
}
\usage{
plusmaze
}
\description{
Data from Gallistel et al. (2004). Original data description: "These are
frequencies, so, in principle, a chisquare test is appropriate to test for changes in
the expected frequency of correct choice. However, for some frequency data sets, the
chisquare test runs into computational difficulties. The chisquare computation is not
valid unless the expected number of observations in the cell with the smallest
expectations is 5 or greater. The function tests for this, and, if this condition is
not met, then it uses the Fisher's exact test. However, Fisher's exact test uses
factorials, and these can become intractably large. This happens in data where the
frequency before a change is already high, say 0.85, and it becomes even higher after
the change. Under these conditions, the numbers of observations in the more populous
cells become large and the factorials intractable. One can also treat these data as
generated by a random rate process that has a certain probability of generating a
correct choice on any given trial. In that case, one would use the random rate test.
This test is more computationally robust (less likely to run afoul of computational
problems) than the chi square test and should be used whenever the chi square test
fails for computational reasons."
}
\keyword{datasets}
