% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logits.R
\name{rrd}
\alias{rrd}
\title{Compute the pseudologit vector for random rate discrete data, using binomial test}
\usage{
rrd(Cum, R)
}
\arguments{
\item{Cum}{Input vector of cumulative response counts}

\item{R}{A vector with trial numbers of putative change points}
}
\value{
L A vector giving for each trial the pseudologit, approximately
the log of the odds that there has been a change point
}
\description{
Compute the pseudologit vector for random rate discrete data, using binomial test
}
\details{
Not normally called directly, but via the cp_wrapper function instead
}

