% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpfinders.R
\name{trun}
\alias{trun}
\title{Truncate the cumulative record at significant change point}
\usage{
trun(Cum, R, L, Crit)
}
\arguments{
\item{Cum}{Input vector of cumulative response counts}

\item{R}{A vector with trial numbers of putative change points}

\item{L}{A vector giving for each trial the pseudologit, approximately
the log of the odds that there has been a change point}

\item{Crit}{A real-valued decision criterion on logit}
}
\value{
A list of Cumt, the truncated cumulative record,
Lt, the truncated L vector up to the row at which a change point was detected,
and r, the row at which L was truncated.
}
\description{
Truncate the cumulative record at significant change point
}
\details{
Not normally called directly, but via the cp_wrapper function instead.
Works for both discrete and continuous cumulative records. All input arguments are
obligatory.
}
