% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdetectoR.R
\docType{data}
\name{watermaze}
\alias{watermaze}
\title{Data on swim efficiencies in a rat learning a water maze}
\format{An object of class \code{data.frame} with 33 rows and 1 columns.}
\source{
\url{http://www.pnas.org/content/suppl/2004/08/31/0404965101.DC1/04965DataSet6.txt/}
}
\usage{
watermaze
}
\description{
Data from Gallistel et al. (2004). Original data description: "The swim efficiency is
the straight-line distance between where the rat is placed in the tank and the location
of the platform divided by the distance actually swum in reaching the platform. These
are again discrete-trial data. The measures can fall anywhere in the interval from 0 to
1. They do not appear to be normally distributed, so one might want to use the K-S
statistic. However, there are only 32 trials (data) and the K-S test requires a minimum
of 4 data in each sample (before and after a putative change point), so one might also
want to try the t test on these data."
}
\keyword{datasets}
